# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Kebutuhan untuk Portal Teknologi",
        "title": "API Teknologi",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
        "/api/getdata-effisiensi_amoniak": {
            "get": {
                "description": "Mengambil data Efisiensi Amoniak",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Efisiensi Dailies"
                ],
                "summary": "Get Data Efisiensi Amoniak",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalSuccessHandlerResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalErrorHandlerResp"
                        }
                    }
                }
            }
        },
        "/api/getdata-effisiensi_urea": {
            "get": {
                "description": "Mengambil data Efisiensi Urea",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Efisiensi Dailies"
                ],
                "summary": "Get Data Efisiensi Urea",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalSuccessHandlerResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalErrorHandlerResp"
                        }
                    }
                }
            }
        },
        "/api/getdata-effisiensi_utilitas": {
            "get": {
                "description": "Mengambil data Efisiensi Utilitas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Efisiensi Dailies"
                ],
                "summary": "Get Data Efisiensi Utilitas",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalSuccessHandlerResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalErrorHandlerResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.GlobalErrorHandlerResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "models.GlobalSuccessHandlerResp": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {},
                "message": {},
                "page": {},
                "status": {
                    "type": "boolean"
                },
                "total": {}
            }
        },
...
    }
}